requirements.txt
matplotlib==3.3.4
pandas==1.1.5
python-box==6.0.2
scipy==1.4.1
tensorflow==2.2.0
tensorflow-probability==0.10.1
tqdm==4.64.0